Add Params.Int()
