+++ taskflow.go
@@ -1,19 +1,3 @@
-/*
-Package taskflow helps implementing build automation.
-It is intended to be used in concert with the "go run" command,
-to run a program which implements the build pipeline (called taskflow).
-A taskflow consists of a set of registered tasks.
-A task has a name, can have a defined command, which is a function with signature
-	func (*taskflow.TF)
-and can have dependencies (already defined tasks).
-
-When the taskflow is executed for given tasks,
-then the tasks' commands are run in the order defined by their dependencies.
-The task's dependencies are run in a recusrive manner, however each is going to be run at most once.
-
-The taskflow is interupted in case a command fails.
-Within these functions, use the Error, Fail or related methods to signal failure.
-*/
 package taskflow
 
 import (
@@ -53,10 +37,6 @@ type Taskflow struct {
 type RegisteredTask struct {
 	name string
 }
-
-// Params represents Taskflow parameters used within Taskflow.
-// The default values set in the struct are overridden in Run method.
-type Params map[string]string
 
 // New return a valid instance of Taskflow with DefaultOutput and initized Params.
 func New() *Taskflow {
